+++ src/socketlib/basic/buffer.py
@@ -12,10 +12,7 @@ class Buffer:
         """ Get a message from the socket until the end of message is reached.
         """
         while msg_end not in self.buffer:
-            try:
-                data = self.socket.recv(1024)
-            except socket.timeout:
-                return
+            data = self.socket.recv(1024)
             if not data:  # socket closed
                 return None
             self.buffer += data