+++ src/socketlib/server/server.py
@@ -3,7 +3,7 @@ import queue
 import logging
 import socket
 import threading
-from typing import Callable, Optional
+from typing import Callable, Optional, Type
 
 from socketlib.basic.buffer import Buffer
 from socketlib.basic.receive import receive_and_enqueue
@@ -229,7 +229,7 @@ class ServerSender(ServerBase):
     def __init__(
             self,
             address: tuple[str, int],
-            to_send: Optional[queue.Queue[str]] = None,
+            to_send: Optional[queue.Queue[str | bytes]] = None,
             reconnect: bool = True,
             timeout: Optional[float] = None,
             stop: Optional[Callable[[], bool]] = None,
@@ -258,9 +258,10 @@ class ServerSender(ServerBase):
         self._run_thread = threading.Thread(
             target=self._send, daemon=True
         )
+        self.send_wait = 0
 
     @property
-    def to_send(self) -> queue.Queue[str]:
+    def to_send(self) -> queue.Queue[str | bytes]:
         return self._to_send
 
     @property
@@ -292,7 +293,8 @@ class ServerSender(ServerBase):
                     timeout=self._timeout,
                     logger=self._logger,
                     name=self.__class__.__name__,
-                    encoding=self.encoding
+                    encoding=self.encoding,
+                    wait=self.send_wait
                 )
                 self.close_connection()
                 self.listen()
@@ -306,7 +308,8 @@ class ServerSender(ServerBase):
                 timeout=self._timeout,
                 logger=self._logger,
                 name=self.__class__.__name__,
-                encoding=self.encoding
+                encoding=self.encoding,
+                wait=self.send_wait
             )
 
 
@@ -320,7 +323,7 @@ class Server(ServerBase):
             self,
             address: tuple[str, int],
             received: Optional[queue.Queue[bytes]] = None,
-            to_send: Optional[queue.Queue[str]] = None,
+            to_send: Optional[queue.Queue[str | bytes]] = None,
             reconnect: bool = True,
             timeout: Optional[float] = None,
             stop_receive: Optional[Callable[[], bool]] = None,
@@ -362,6 +365,8 @@ class Server(ServerBase):
         self._recv_thread = threading.Thread(target=self._recv, daemon=True)
         self._connected = threading.Event()
 
+        self.send_wait = 0
+
     @property
     def to_send(self) -> queue.Queue[str]:
         return self._to_send
@@ -390,7 +395,8 @@ class Server(ServerBase):
                     timeout=self._timeout,
                     logger=self._logger,
                     name=self.__class__.__name__,
-                    encoding=self.encoding
+                    encoding=self.encoding,
+                    wait=self.send_wait
                 )
                 self._connected.clear()
                 self.close_connection()
@@ -405,7 +411,8 @@ class Server(ServerBase):
                 timeout=self._timeout,
                 logger=self._logger,
                 name=self.__class__.__name__,
-                encoding=self.encoding
+                encoding=self.encoding,
+                wait=self.send_wait
             )
 
     def _recv(self):